Move resolving type variables to model core
