+++ subprojects/model-core/src/main/java/org/gradle/internal/reflect/JavaPropertyReflectionUtil.java
@@ -16,14 +16,24 @@
 
 package org.gradle.internal.reflect;
 
+import com.google.common.reflect.TypeToken;
 import org.apache.commons.lang.reflect.MethodUtils;
 import org.gradle.internal.UncheckedException;
+import org.gradle.util.CollectionUtils;
 
 import javax.annotation.Nullable;
 import java.lang.annotation.Annotation;
 import java.lang.annotation.Inherited;
+import java.lang.reflect.GenericArrayType;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.lang.reflect.TypeVariable;
+import java.lang.reflect.WildcardType;
+import java.util.ArrayDeque;
+import java.util.List;
+import java.util.Queue;
 import java.util.Set;
 import java.util.WeakHashMap;
 
@@ -150,6 +160,116 @@ public class JavaPropertyReflectionUtil {
         } catch (java.lang.NoSuchMethodException e) {
             throw UncheckedException.throwAsUncheckedException(e);
         }
+    }
+
+    /**
+     * Resolves the return type of a method in a given class.
+     *
+     * For example, for {@code MyList implements List<String>}, resolving the return type of {@link List#get(int)} in {@code MyList} yields {@link String}.
+     */
+    public static <T> Type resolveMethodReturnType(Class<T> type, Method method) {
+        Type returnType = method.getGenericReturnType();
+        if (type.equals(method.getDeclaringClass())) {
+            // No need to resolve type parameters if the method is from the same class.
+            return returnType;
+        }
+        // Checking if there is a type variable to resolve, since resolving the type variable via `TypeToken` is quite expensive.
+        return hasTypeVariable(returnType) ? TypeToken.of(type).method(method).getReturnType().getType() : returnType;
+    }
+
+    /**
+     * Checks if a type has a type variable which may require resolving.
+     */
+    public static boolean hasTypeVariable(Type type) {
+        // do some checks up-front, so we avoid creating the queue in most cases
+        // Cases we want to handle:
+        // - List<String>
+        // - Class<?>
+        // - List<Class<?>>
+        // - Integer[]
+        // - ? extends BaseType
+        // - Class<?>[]
+        if (doesNotHaveTypeVariable(type)) {
+            return false;
+        }
+        if (type instanceof TypeVariable) {
+            return true;
+        }
+        if (type instanceof ParameterizedType) {
+            ParameterizedType parameterizedType = (ParameterizedType) type;
+            boolean noTypeVariables = true;
+            for (Type actualTypeArgument : parameterizedType.getActualTypeArguments()) {
+                if (actualTypeArgument instanceof TypeVariable) {
+                    return true;
+                }
+                noTypeVariables &= doesNotHaveTypeVariable(actualTypeArgument);
+            }
+            if (noTypeVariables) {
+                return false;
+            }
+        }
+        if (type instanceof GenericArrayType) {
+            GenericArrayType genericArrayType = (GenericArrayType) type;
+            if (genericArrayType.getGenericComponentType() instanceof TypeVariable) {
+                return true;
+            }
+        }
+
+        // Type is more complicated, need to check everything.
+        Queue<Type> typesToInspect = new ArrayDeque<Type>();
+        typesToInspect.add(type);
+        while (!typesToInspect.isEmpty()) {
+            Type typeToInspect = typesToInspect.remove();
+            if (typeToInspect instanceof Class) {
+                continue;
+            }
+            if (typeToInspect instanceof TypeVariable) {
+                return true;
+            }
+            if (typeToInspect instanceof ParameterizedType) {
+                ParameterizedType parameterizedType = (ParameterizedType) typeToInspect;
+                CollectionUtils.addAll(typesToInspect, parameterizedType.getActualTypeArguments());
+            } else if (typeToInspect instanceof GenericArrayType) {
+                GenericArrayType arrayType = (GenericArrayType) typeToInspect;
+                typesToInspect.add(arrayType.getGenericComponentType());
+            } else if (typeToInspect instanceof WildcardType) {
+                WildcardType wildcardType = (WildcardType) typeToInspect;
+                CollectionUtils.addAll(typesToInspect, wildcardType.getLowerBounds());
+                CollectionUtils.addAll(typesToInspect, wildcardType.getUpperBounds());
+            } else {
+                // We don't know what the type is - let Guava take care of it.
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Quick check if a type does not have any type variables.
+     *
+     * Handled cases:
+     * - raw Class
+     * - Wildcard type with Class bounds, e.g. ? extends BaseType
+     */
+    private static boolean doesNotHaveTypeVariable(Type type) {
+        if (type instanceof Class) {
+            return true;
+        }
+        if (type instanceof WildcardType) {
+            WildcardType wildcardType = (WildcardType) type;
+            for (Type lowerBound : wildcardType.getLowerBounds()) {
+                if (!(lowerBound instanceof Class)) {
+                    return false;
+                }
+            }
+            for (Type upperBound : wildcardType.getUpperBounds()) {
+                if (!(upperBound instanceof Class)) {
+                    return false;
+                }
+            }
+            return true;
+        }
+        return false;
     }
 
     private static class GetterMethodBackedPropertyAccessor<T, F> implements PropertyAccessor<T, F> {
